add support for term_id when adding/updating annotation

+++ dbbact_jobs/update_ontology_parents.py
@@ -15,7 +15,7 @@ from dbbact_server.utils import debug, SetDebugLevel
 __version__ = "0.9"
 
 
-def update_ontology_parents(con, cur, overwrite=True):
+def update_ontology_parents(con, cur, overwrite=True, ontology=None):
 	'''
 	Fill the database AnnotationParentsTable
 
@@ -26,21 +26,28 @@ def update_ontology_parents(con, cur, overwrite=True):
 
 	overwrite : bool (optional)
 		False (default) to not overwrite existing annotation parents, True to delete all
+	ontology: str or None (optional)
+		if None, update all ontologies. Otherwise, update only terms from the given ontology
 	'''
+	if ontology is not None:
+		raise ValueError('ontology specific operation not supported yet!')
 	skipped = 0
 	added = 0
 	all_parents_dict = {}
 	if overwrite:
-		debug(4, 'deleting old parent counts')
-		# delete the current counts since we are updating all entries (and addparents adds 1 to the counts...)
-		# another way to do it (faster in postgres 11)?
-		# alter table ontologytable drop column seqcount;
-		# alter table ontologytable drop column annotationcount;
-		# alter table ontologytable ADD COLUMN seqCount integer DEFAULT 0;
-		# alter table ontologytable ADD COLUMN annotationCount integer DEFAULT 0;
-		cur.execute('UPDATE OntologyTable SET seqCount=0, annotationCount=0')
-		debug(4, 'deleting annotationparentstable')
-		cur.execute('DELETE FROM AnnotationParentsTable')
+		if ontology is None:
+			debug(4, 'deleting old parent counts')
+			# delete the current counts since we are updating all entries (and addparents adds 1 to the counts...)
+			# another way to do it (faster in postgres 11)?
+			# alter table ontologytable drop column seqcount;
+			# alter table ontologytable drop column annotationcount;
+			# alter table ontologytable ADD COLUMN seqCount integer DEFAULT 0;
+			# alter table ontologytable ADD COLUMN annotationCount integer DEFAULT 0;
+			cur.execute('UPDATE OntologyTable SET seqCount=0, annotationCount=0')
+			debug(4, 'deleting annotationparentstable')
+			cur.execute('DELETE FROM AnnotationParentsTable')
+
+	# iterate over all annotations
 	cur.execute('SELECT id,seqCount from AnnotationsTable')
 	annotations = cur.fetchall()
 	debug(4, 'updating AnnotationParentsTable for %d annotations' % len(annotations))
@@ -51,12 +58,14 @@ def update_ontology_parents(con, cur, overwrite=True):
 		cseqcount = cres[1]
 		if cseqcount == 0:
 			debug(5, 'WARNING: annotation %d has no sequences in AnnotationsTable' % cid)
+
 		# if not in overwrite mode, don't add parents to entries already in the table
 		if not overwrite:
 			cur.execute('SELECT idAnnotation from AnnotationParentsTable WHERE idAnnotation=%s', [cid])
 			if cur.rowcount > 0:
 				skipped += 1
 				continue
+
 		err, annotationdetails = dbannotations.get_annotation_details_termids(con, cur, cid)
 		if err:
 			debug(6, 'error: %s' % err)
@@ -78,6 +87,7 @@ def main(argv):
 	parser.add_argument('--password', help='postgres password', default='magNiv')
 	parser.add_argument('--proc-title', help='name of the process (to view in ps aux)')
 	parser.add_argument('--debug-level', help='debug level (1 for debug ... 9 for critical)', default=2, type=int)
+	parser.add_argument('--ontology', help='ontology to update parents for (empty to update all')
 	args = parser.parse_args(argv)
 
 	SetDebugLevel(args.debug_level)
@@ -86,7 +96,7 @@ def main(argv):
 		setproctitle.setproctitle(args.proc_title)
 
 	con, cur = db_access.connect_db(database=args.database, user=args.user, password=args.password, port=args.port, host=args.host)
-	update_ontology_parents(con, cur)
+	update_ontology_parents(con, cur, ontology=args.ontology)
 
 
 if __name__ == "__main__":
